Add feature to create new grid with user input
Added another feature to the clear button that allows the user to input
a number, which determines the number of squares per side for a newly
created grid. This new feature conflicted with the code that "cleared"
the grid, so the coding for clearing the grid was changed and turned
into the removeGrid() function. Also, createGrid() now accepts a
parameter, and the coding was rearranged for cleaner code.