Add 'change to black' button
Added another button that changes the mouseover function so that the
grid cells change to black. This required changing
addGridEventListener() to accept a parameter that keeps track of which
event listener is currently attached to the grid cells. Now the user can
switch between changing the grid cells to black, or to random RGB
colours.